+++ src/app/account/account.module.scss
@@ -1,8 +1,13 @@
 @import "@styles/vars.scss";
 @import "@styles/mixins.scss";
 
+.auth {
+  height: 100vh;
+  background-color: $gray-100;
+}
+
 .account {
-  margin: 0 auto;
+  margin: 0 auto 200px;
   padding: 32px 16px;
   cursor: default;
 
@@ -79,8 +84,8 @@
   }
 
   .login {
-    margin-left: 8px;
     color: $system-error;
+    text-decoration: underline;
     cursor: pointer;
   }
 }
@@ -123,6 +128,17 @@
   cursor: pointer;
 }
 
+.dataContainer {
+  @include set-flex(column, flex-start, flex-start);
+  width: 100%;
+  gap: 8px;
+  color: $gray-30;
+  font-family: $font-text;
+  font-size: $font-size-md;
+  font-weight: $font-weight-regular;
+  line-height: 20px;
+}
+
 .address {
   cursor: default;
 }
@@ -132,7 +148,7 @@
   height: 48px;
   padding: 12px 16px;
   background-color: $gray-90;
-  color: $gray-30;
+  color: $gray-0;
   font-family: $font-text;
   font-size: $font-size-base;
   line-height: $line-height-base;
@@ -144,6 +160,18 @@
 .input {
   position: relative;
   width: 100%;
+}
+
+.inputData{
+  width: 100%;
+  height: 48px;
+  padding: 12px 16px;
+  font-family: $font-text;
+  font-size: $font-size-base;
+  line-height: $line-height-base;
+  font-weight: $font-weight-regular;
+  text-align: start;
+  color: $gray-0;
 }
 
 .error {
@@ -203,6 +231,27 @@
   }
 }
 
+.signOut {
+  @include set-flex(row, center, center);
+  gap: 10px;
+  height: 48px;
+  width: 100%;
+  padding: 12px 24px;
+  border-radius: 24px;
+  border: 1px solid $gray-5;
+  background-color: transparent;
+  color: $gray-5;
+  font-family: $font-text;
+  font-size: $font-size-base;
+  line-height: $line-height-base;
+  font-weight: $font-weight-regular;
+  text-align: center;
+
+  &:hover {
+    cursor: pointer;
+  }
+}
+
 .accordion {
   margin: 0 auto;
   padding: 0 14px;
@@ -227,7 +276,6 @@
     padding: 22px 0 0;
   }
 }
-
 
 .accordionTitle {
   font-family: $font-text;
