+++ src/app/account/components/registers-form/registers.scss
@@ -1,0 +1,3 @@
+.phoneWrapper {
+  margin: 0 !important;
+}
