+++ src/app/components/collection/collection.module.scss
@@ -2,28 +2,52 @@
 @import "@styles/mixins.scss";
 
 .titleContainer {
-    padding: 22px;
-    width: 100%;
+	padding: 22px;
+	width: 100%;
+	@include set-flex(column, center, center);
 
-    @include tabletXL {
-        padding: 36px;
-    }
+	@include tabletXL {
+			padding: 36px;
+	}
+}
 
-    .title {
-        color: $gray-5;
-        font-family: $font-title;
-        font-weight: $font-weight-regular;
-        font-size: $font-size-xxl;
-        line-height: $line-height-xxl;
+.title {
+	margin-bottom: 24px;
+	color: $gray-5;
+	font-family: $font-title;
+	font-weight: $font-weight-regular;
+	font-size: $font-size-xl;
+	line-height: $line-height-xxl;
 
-        @include tabletXL {
-            font-size: 48px;
-            line-height: 56px;
-        }
+	@include tabletXL {
+		font-size: 48px;
+		line-height: 56px;
+	}
 
-        @include desktop {
-            font-size: $font-size-huge;
-            line-height: $line-height-huge;
-        }
-    }
+	@include desktop {
+		font-size: $font-size-huge;
+		line-height: $line-height-huge;
+	}
+}
+
+.button {
+	width: 316px;
+	border-radius: 24px;
+	border: 1px solid $gray-5;
+	display: flex;
+	padding: 12px 24px;
+	justify-content: center;
+	align-items: center;
+	background-color: transparent;
+	color: $gray-5;
+	font-size: $font-size-base;
+	cursor: pointer;
+
+	@include tabletXL {
+			max-width: 640px;
+	}
+
+	@include laptop {
+			max-width: 100%;
+	}
 }
