+++ src/app/components/collection/collection.tsx
@@ -1,12 +1,22 @@
-import React from "react";
+import React, { FC } from "react";
 import style from "./collection.module.scss";
+import Link from "next/link";
 
-const Collection = () => {
+type Props = {
+  title: string;
+  isArtsTab?: boolean;
+};
+
+const Collection: FC<Props> = ({ title, isArtsTab }) => {
     return (
         <div className={style.titleContainer}>
             <div className={style.title}>
-                This feature is under development now!
+                {title}
             </div>
+
+            {isArtsTab && (
+              <Link href="/profile/createPainting" className={style.button}>Add arts</Link>
+            )}
         </div>
     );
 };
