+++ src/app/icons/icon-logout.tsx
@@ -1,0 +1,14 @@
+export const LogoutIcon = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="25"
+    height="24"
+    viewBox="0 0 25 24"
+    fill="none"
+  >
+    <path
+      d="M10.8125 20.25C10.8125 20.3992 10.7532 20.5423 10.6477 20.6477C10.5423 20.7532 10.3992 20.8125 10.25 20.8125H5C4.6519 20.8125 4.31806 20.6742 4.07192 20.4281C3.82578 20.1819 3.6875 19.8481 3.6875 19.5V4.5C3.6875 4.1519 3.82578 3.81806 4.07192 3.57192C4.31806 3.32578 4.6519 3.1875 5 3.1875H10.25C10.3992 3.1875 10.5423 3.24676 10.6477 3.35225C10.7532 3.45774 10.8125 3.60082 10.8125 3.75C10.8125 3.89918 10.7532 4.04226 10.6477 4.14775C10.5423 4.25324 10.3992 4.3125 10.25 4.3125H5C4.95027 4.3125 4.90258 4.33225 4.86742 4.36742C4.83225 4.40258 4.8125 4.45027 4.8125 4.5V19.5C4.8125 19.5497 4.83225 19.5974 4.86742 19.6326C4.90258 19.6677 4.95027 19.6875 5 19.6875H10.25C10.3992 19.6875 10.5423 19.7468 10.6477 19.8523C10.7532 19.9577 10.8125 20.1008 10.8125 20.25ZM21.1475 11.6025L17.3975 7.8525C17.2909 7.75314 17.1498 7.69905 17.0041 7.70162C16.8584 7.70419 16.7193 7.76322 16.6163 7.86628C16.5132 7.96934 16.4542 8.10838 16.4516 8.25411C16.449 8.39983 16.5031 8.54087 16.6025 8.6475L19.3916 11.4375H10.25C10.1008 11.4375 9.95774 11.4968 9.85225 11.6023C9.74676 11.7077 9.6875 11.8508 9.6875 12C9.6875 12.1492 9.74676 12.2923 9.85225 12.3977C9.95774 12.5032 10.1008 12.5625 10.25 12.5625H19.3916L16.6025 15.3525C16.5472 15.404 16.5029 15.4661 16.4722 15.5351C16.4414 15.6041 16.4249 15.6786 16.4236 15.7541C16.4222 15.8296 16.4361 15.9047 16.4644 15.9747C16.4927 16.0447 16.5348 16.1084 16.5882 16.1618C16.6416 16.2152 16.7053 16.2573 16.7753 16.2856C16.8453 16.3139 16.9204 16.3278 16.9959 16.3264C17.0714 16.3251 17.1459 16.3086 17.2149 16.2778C17.2839 16.2471 17.346 16.2028 17.3975 16.1475L21.1475 12.3975C21.2528 12.292 21.312 12.1491 21.312 12C21.312 11.8509 21.2528 11.708 21.1475 11.6025Z"
+      fill="#EFF0F1"
+    />
+  </svg>
+);
