+++ src/app/profile/page.module.scss
@@ -1,7 +1,6 @@
 @import "@styles/vars.scss";
 
 .auth {
-  // padding: 100px 0;
   height: 100vh;
   background-color: $gray-100;
 }
