+++ src/utils/api.ts
@@ -370,11 +370,44 @@ export async function getStripeLink(
   return response.data;
 }
 
+export async function getAccount(headers: Headers) {
+  const { data } = await axios.get(
+    `${BASE_URL}account`,
+    {
+      headers: {
+        ...headers,
+        "Content-Type": "application/json;charset=utf-8",
+      },
+    }
+  );
+
+  return data;
+};
+
+
 export async function createAccount(
   headers: Headers,
   accountData: AccountData,
 ) {
   const { data } = await axios.post(
+    `${BASE_URL}account`,
+    accountData,
+    {
+      headers: {
+        ...headers,
+        "Content-Type": "application/json;charset=utf-8",
+      },
+    }
+  );
+
+  return data;
+};
+
+export async function updateAccount(
+  headers: Headers,
+  accountData: AccountData,
+) {
+  const { data } = await axios.put(
     `${BASE_URL}account`,
     accountData,
     {
